feat: Implement format currency

+++ internal/domain/user.go
@@ -10,7 +10,7 @@ type User struct {
 
 	Name     string          `json:"name" validate:"required"`
 	Currency string          `json:"currency" validate:"required"`
-	Scale    uint32          `json:"scale" validate:"required"`
+	Scale    uint8           `json:"scale" validate:"required"`
 	Balance  decimal.Decimal `json:"balance" validate:"required"`
 
 	CreatedDate time.Time `json:"createdDate"`
