+++ shared/dto/user/user.go
@@ -7,6 +7,6 @@ type UserDto struct {
 
 	Name     string          `json:"name" validate:"required"`
 	Currency string          `json:"currency" validate:"required"`
-	Scale    uint32          `json:"scale" validate:"required"`
+	Scale    uint8           `json:"scale" validate:"required"`
 	Balance  decimal.Decimal `json:"balance" validate:"required"`
 }
